% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integrate.R
\name{data_integrate}
\alias{data_integrate}
\title{Integrate data for MagicalRsq model training and testing}
\usage{
data_integrate(
  file,
  header = T,
  chr,
  pos_col,
  rr_dir = NULL,
  SHIC_dir = NULL,
  AFLD_dir = NULL,
  outfile = NULL
)
}
\arguments{
\item{file}{Required. Input file name. The file should be chromosom-specific, i.e., all the variants should be on the same chromosome.
It should contain minimum of position (in hg38), reference allele and alternative allele, with column names as POS, REF and ALT respectively.
Strongly recommended to include Rsq from minimac (or info from IMPUTE) given by any imputation software to improve model performance.
The file could also contain true R2 information for model training or evaluation purpose.}

\item{header}{Logical. The parameter is passing to the \code{fread} function. TRUE by default. Strongly recommended containing a header in your data file.}

\item{chr}{Chromosome number. Required, integer.}

\item{pos_col}{GRCH38/hg38 position column number. Required, integer.}

\item{rr_dir}{Directory containing downloaded recombination rate from 1000G. Optional.}

\item{SHIC_dir}{Directory containing downloaded SHIC features. Optional.}

\item{AFLD_dir}{Directory containing downloaded population-specifc allele frequencies and LD scores. Optional.}

\item{outfile}{Output file name. Optional but strongly recommended.}
}
\value{
Dataframe of integrated data that could be used as MagicalRsq training and testing input.
}
\description{
This function is used to integrate user-input dataset containing Rsq information with provided recombination rate, SHIC features, allele frequencies, and LD scores in different populations.
Recombination rate can be downloaded at ftp://yunlianon:anon@rc-ns-ftp.its.unc.edu/MagicalRsqX/RecombRate/.
SHIC features can be downloaded at ftp://yunlianon:anon@rc-ns-ftp.its.unc.edu/MagicalRsq/SHIC/.
Population-specific allele frequencies and LD scores can be downloaded at ftp://yunlianon:anon@rc-ns-ftp.its.unc.edu/MagicalRsqX/AFLD/.
}
\examples{

toy_chr22 = data_integrate("toy_chr22_50k.txt.gz", chr = 22, 
pos_col = 2, SHIC_dir = "../SHIC/", AFLD_dir = "../AFLD/",
rr_dir = "../RecombRate/", outfile = "toy_chr22_integrated.txt.gz")

}
